Moved back UserActivity.py for rebase
